fix: Convert coordinates to int for PyQt5 drawLine compatibility
PyQt5's drawLine() has multiple overloads but none accept 4 separate float args:
- drawLine(int, int, int, int) ✓
- drawLine(QLineF)
- drawLine(QPointF, QPointF)

Converting to int instead of float matches the correct overload signature.
Since we're drawing on pixel coordinates, integer conversion is appropriate.

Error was:
  TypeError: arguments did not match any overloaded call:
  drawLine(self, x1: int, y1: int, x2: int, y2: int): argument 1 has unexpected type 'float'

+++ image_velocimetry_tools/gui/stiv_processor.py
@@ -922,8 +922,8 @@ def draw_line_on_pixmap(original_pixmap, angle_degrees=np.nan, color=Qt.yellow):
     start_y = center_y - dy
 
     # Draw the line on the modified QPixmap
-    # Convert numpy types to Python floats for PyQt5 compatibility
-    painter.drawLine(float(start_x), float(start_y), float(end_x), float(end_y))
+    # Convert to integers for PyQt5's drawLine(int, int, int, int) overload
+    painter.drawLine(int(start_x), int(start_y), int(end_x), int(end_y))
 
     # End painting
     painter.end()
